fix and improve
